requirements.txt
streamlit==1.27.2
face-alignment==1.3.5
torch==1.13.1+cpu
torchvision==0.14.1+cpu
opencv-python==4.5.5.64
numpy==1.23.5
Pillow==9.5.0
gspread==5.11.1
oauth2client==4.1.3
imageio[ffmpeg]